Prompt user to click after games loads to ensure music starts
Some browsers now mute a page's audio until a user has made a gesture. The
game now prompts the user to click before loading the menu screen. This
ensures that the menu music plays immediately. Previously, the menu audio was
muted when the menu was shown for the first time. The audio did not begin
until the user clicked start.